% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{encontrar_pior_ano_time}
\alias{encontrar_pior_ano_time}
\title{Obtem o ano em que um time obteve pior desempenho}
\usage{
encontrar_pior_ano_time(time)
}
\arguments{
\item{time}{Um vetor de caracteres informando o nome do time}
}
\value{
retorna um data frame contedo tres colunas
}
\description{
Obtem o ano em que um time obteve pior desempenho
}
\examples{
encontrar_pior_ano_time("Flamengo")
}
