% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{partidas_brasileirao}
\alias{partidas_brasileirao}
\title{Resultado do campeonato brasileiro}
\format{
Um spec_tbl_df com 266 linhas e 7 colunas:
\describe{
\item{temporada}{Ano em que aconteceu o jogo}
\item{data}{Data em que aconteu o jogo}
\item{time_casa}{Time da casa}
\item{gols}{Placar do jogo}
\item{time_visitante}{Time visitante}
\item{gols_casa}{Gols marcados pelo time da casa}
\item{gols_visitante}{Gols marcados pelo visitante}
\item{quem_ganhou}{Quem ganhou a partida}
}
}
\source{
\url{"https://git.io/JOqUN"}
}
\usage{
partidas_brasileirao
}
\description{
Resultado dos jogos do campeonato brasileiro
}
\keyword{datasets}
